testfile up again
me and a few others created an attempt to sideload our app onto those device. 
BUT virustotal and other sites display a backdoor trojan.

(I do not know if thats real. I DO NOT RECOMMEND DOWNLOADING)
Im not responsible for ANY of your devices.

This is just for me to log our progress, i will watch and figure out why it says backdoors are in it.
until i figure it out, i do not recommend downloading.